geekle_ia_models/requirements-llama.txt
--index-url https://abetlen.github.io/llama-cpp-python/whl/cu121
llama-cpp-python==0.2.82